+++ s3torchconnectorclient/rust/src/python_structs.rs
@@ -6,3 +6,4 @@
 pub(crate) mod py_list_object_result;
 pub(crate) mod py_object_info;
 pub(crate) mod py_restore_status;
+pub(crate) mod py_head_object_result;